requirements.txt
--extra-index-url https://download.pytorch.org/whl/cu121

fastapi
uvicorn
httpx
openai-whisper
elevenlabs
requests
pydantic
aiofiles
python-multipart
psycopg2-binary
pyyaml
pyjwt
transformers
ctransformers
accelerate
huggingface-hub
langchain
langchain-core
langchain-community
langchain-huggingface
elevenlabs
langchain-openai
langchain-anthropic
langgraph
langsmith
openai
tavily-python
python-dotenv
deepl
nemoguardrails

# Version freeze with CUDA 12.1
torch==2.2.0+cu121
torchvision
torchaudio==2.2.0+cu121

